backend/requirements.txt
fastapi
uvicorn
httpx
pydantic
python-dotenv
azure-cosmos
python-jose[cryptography]
multipart
passlib[argon2]
email-validator